refactor: change enum's parameter and move the strings to strings file
