--- design_system/src/main/java/com/moscow/cineverse/designSystem/component/ExploreTabsPages.kt
@@ -1,7 +1,0 @@
-package com.moscow.cineverse.designSystem.component
-
-enum class ExploreTabsPages(val text: String) {
-    MOVIES("Movies"),
-    SERIES("Series"),
-    ACTORS("Actors")
-}
