+++ design_system/src/main/java/com/moscow/cineverse/designSystem/component/tabs/ExploreTabsPages.kt
@@ -1,0 +1,9 @@
+package com.moscow.cineverse.designSystem.component.tabs
+
+import com.example.design_system.R
+
+enum class ExploreTabsPages(val textId: Int) {
+    MOVIES(R.string.movies),
+    SERIES(R.string.series),
+    ACTORS(R.string.actors)
+}
